+++ frontend/src/app/components/game-delete/game-delete.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-game-delete',
+  imports: [],
+  templateUrl: './game-delete.component.html',
+  styleUrl: './game-delete.component.css'
+})
+export class GameDeleteComponent {
+
+}