+++ frontend/CHANGELOG.md
@@ -5,14 +5,20 @@
 - Test of changeLog
 - Creation of Login and Welcome Page
 
-## [v0.0.1] - 2025-05-27 
+## [v0.1.0] - 2025-05-27 
 ### WIP Version
 - Admin menus
 
-## [v0.0.2] - 2025-05-30 
+## [v0.2.0] - 2025-05-30 
 ### WIP Version
 - Admin staff modify is created
 - Staff login works
+
+## [v0.3.0] - 2025-05-30 
+### WIP Version
+- Staff menu works
+- Can add, list and search games
+- Changed versions